+++ src/main/java/org/folio/search/integration/KafkaMessageListener.java
@@ -4,6 +4,7 @@ import static org.apache.commons.collections4.MapUtils.getString;
 import static org.apache.commons.lang3.RegExUtils.replaceAll;
 import static org.folio.search.configuration.RetryTemplateConfiguration.KAFKA_RETRY_TEMPLATE_NAME;
 import static org.folio.search.domain.dto.ResourceEventType.CREATE;
+import static org.folio.search.domain.dto.ResourceEventType.DELETE;
 import static org.folio.search.domain.dto.ResourceEventType.REINDEX;
 import static org.folio.search.utils.SearchConverterUtils.getEventPayload;
 import static org.folio.search.utils.SearchConverterUtils.getResourceEventId;
@@ -27,7 +28,9 @@ import org.apache.kafka.clients.consumer.ConsumerRecord;
 import org.apache.logging.log4j.message.FormattedMessage;
 import org.folio.search.domain.dto.ResourceEvent;
 import org.folio.search.model.event.ConsortiumInstanceEvent;
+import org.folio.search.model.types.ResourceType;
 import org.folio.search.service.ResourceService;
+import org.folio.search.service.config.ConfigSynchronizationService;
 import org.folio.search.utils.KafkaConstants;
 import org.folio.spring.service.SystemUserScopedExecutionService;
 import org.springframework.kafka.annotation.KafkaListener;
@@ -44,6 +47,7 @@ public class KafkaMessageListener {
   private final ResourceService resourceService;
   private final FolioMessageBatchProcessor folioMessageBatchProcessor;
   private final SystemUserScopedExecutionService executionService;
+  private final ConfigSynchronizationService configSynchronizationService;
 
   /**
    * Handles instance events and indexes them by id.
@@ -149,6 +153,28 @@ public class KafkaMessageListener {
     batchByTenant.forEach((tenant, resourceEvents) -> executionService.executeSystemUserScoped(tenant, () -> {
       folioMessageBatchProcessor.consumeBatchWithFallback(batch, KAFKA_RETRY_TEMPLATE_NAME,
         resourceService::indexConsortiumInstances, KafkaMessageListener::logFailedConsortiumEvent);
+      return null;
+    }));
+  }
+
+  @KafkaListener(
+    id = KafkaConstants.CLASSIFICATION_TYPE_LISTENER_ID,
+    containerFactory = "standardListenerContainerFactory",
+    groupId = "#{folioKafkaProperties.listener['classification-type'].groupId}",
+    concurrency = "#{folioKafkaProperties.listener['classification-type'].concurrency}",
+    topicPattern = "#{folioKafkaProperties.listener['classification-type'].topicPattern}")
+  public void handleClassificationTypeEvents(List<ConsumerRecord<String, ResourceEvent>> consumerRecords) {
+    log.info("Processing classification-type events from Kafka [number of events: {}]", consumerRecords.size());
+    var batch = consumerRecords.stream()
+      .map(ConsumerRecord::value)
+      .filter(resourceEvent -> resourceEvent.getType() == DELETE).toList();
+
+    var batchByTenant = batch.stream().collect(Collectors.groupingBy(ResourceEvent::getTenant));
+
+    batchByTenant.forEach((tenant, resourceEvents) -> executionService.executeSystemUserScoped(tenant, () -> {
+      folioMessageBatchProcessor.consumeBatchWithFallback(batch, KAFKA_RETRY_TEMPLATE_NAME,
+        resourceEvent -> configSynchronizationService.sync(resourceEvent, ResourceType.CLASSIFICATION_TYPE),
+        KafkaMessageListener::logFailedEvent);
       return null;
     }));
   }