+++ src/main/java/org/folio/search/utils/KafkaConstants.java
@@ -6,6 +6,7 @@ public final class KafkaConstants {
   public static final String SUBJECT_LISTENER_ID = "mod-search-subject-listener";
   public static final String EVENT_LISTENER_ID = "mod-search-events-listener";
   public static final String CONSORTIUM_INSTANCE_LISTENER_ID = "mod-search-consortium-instance-listener";
+  public static final String CLASSIFICATION_TYPE_LISTENER_ID = "mod-search-classification-type-listener";
 
   private KafkaConstants() {}
 }