+++ src/test/java/org/folio/search/service/config/BrowseConfigServiceTest.java
@@ -1,8 +1,13 @@
 package org.folio.search.service.config;
 
+import static java.util.Collections.emptyList;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.folio.search.client.InventoryReferenceDataClient.ReferenceDataType.CLASSIFICATION_TYPES;
 import static org.folio.search.domain.dto.BrowseOptionType.ALL;
 import static org.folio.search.domain.dto.BrowseOptionType.LC;
 import static org.folio.search.domain.dto.BrowseType.INSTANCE_CLASSIFICATION;
+import static org.folio.search.model.client.CqlQueryParam.ID;
+import static org.folio.search.utils.TestUtils.randomId;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
@@ -10,9 +15,13 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.verifyNoInteractions;
 
+import java.util.HashSet;
 import java.util.List;
 import java.util.Optional;
+import java.util.Set;
+import java.util.UUID;
 import org.folio.search.converter.BrowseConfigMapper;
 import org.folio.search.domain.dto.BrowseConfig;
 import org.folio.search.domain.dto.BrowseConfigCollection;
@@ -20,6 +29,7 @@ import org.folio.search.domain.dto.BrowseOptionType;
 import org.folio.search.domain.dto.BrowseType;
 import org.folio.search.domain.dto.ShelvingOrderAlgorithmType;
 import org.folio.search.exception.RequestValidationException;
+import org.folio.search.integration.ReferenceDataService;
 import org.folio.search.model.config.BrowseConfigEntity;
 import org.folio.search.model.config.BrowseConfigId;
 import org.folio.search.repository.BrowseConfigEntityRepository;
@@ -27,6 +37,7 @@ import org.folio.spring.testing.type.UnitTest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.ArgumentCaptor;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
@@ -39,18 +50,27 @@ class BrowseConfigServiceTest {
   private BrowseConfigEntityRepository repository;
   @Mock
   private BrowseConfigMapper mapper;
+
+  @Mock
+  private ReferenceDataService referenceDataService;
+
   @InjectMocks
   private BrowseConfigService service;
 
   private BrowseType type;
   private BrowseOptionType configId;
   private BrowseConfig config;
+  private List<String> typeIds;
 
   @BeforeEach
   void setUp() {
     type = INSTANCE_CLASSIFICATION;
     configId = LC;
-    config = new BrowseConfig().id(LC).shelvingAlgorithm(ShelvingOrderAlgorithmType.LC);
+    typeIds = List.of(randomId(), randomId());
+    config = new BrowseConfig()
+      .id(LC)
+      .shelvingAlgorithm(ShelvingOrderAlgorithmType.LC)
+      .typeIds(typeIds.stream().map(UUID::fromString).toList());
   }
 
   @Test
@@ -91,9 +111,11 @@ class BrowseConfigServiceTest {
   }
 
   @Test
-  void shouldUpsertConfigWhenConfigIdMatches() {
+  void shouldUpsertConfigWhenFitAllValidations() {
     var entity = getEntity();
     given(mapper.convert(type, config)).willReturn(entity);
+    given(referenceDataService.fetchReferenceData(CLASSIFICATION_TYPES, ID, new HashSet<>(typeIds)))
+      .willReturn(new HashSet<>(typeIds));
 
     assertDoesNotThrow(() -> service.upsertConfig(type, configId, config));
     verify(repository).save(entity);
@@ -105,8 +127,37 @@ class BrowseConfigServiceTest {
 
     var exception = assertThrows(RequestValidationException.class, () -> service.upsertConfig(type, configId, config));
 
-    String expectedMessage = String.format("Body doesn't match path parameter: %s", configId.getValue());
-    assertTrue(exception.getMessage().contains(expectedMessage));
+    assertThat(exception)
+      .hasMessage("Body doesn't match path parameter: %s", configId.getValue());
+    verifyNoInteractions(repository);
+  }
+
+  @Test
+  void shouldThrowExceptionWhenIdIsNotInReferenceData() {
+    given(referenceDataService.fetchReferenceData(CLASSIFICATION_TYPES, ID, new HashSet<>(typeIds)))
+      .willReturn(Set.of(typeIds.get(0)));
+
+    var exception = assertThrows(RequestValidationException.class, () -> service.upsertConfig(type, configId, config));
+
+    assertThat(exception)
+      .hasMessage("Classification type IDs don't exist");
+
+    verifyNoInteractions(repository);
+  }
+
+  @Test
+  void shouldDeleteTypeIdsFromConfigs() {
+    List<BrowseConfigEntity> entities = List.of(getEntity(), getEntity());
+    given(repository.findByConfigId_BrowseType(type.getValue())).willReturn(entities);
+    ArgumentCaptor<List<BrowseConfigEntity>> captor = ArgumentCaptor.captor();
+    given(repository.saveAll(captor.capture())).willReturn(emptyList());
+
+    service.deleteTypeIdsFromConfigs(type, List.of("e1", "e2"));
+
+    var captorValues = captor.getValue();
+    for (BrowseConfigEntity captorValue : captorValues) {
+      assertThat(captorValue.getTypeIds()).isNullOrEmpty();
+    }
   }
 
   private static BrowseConfigEntity getEntity() {